[api] move createAccessTokenForCnav in cnav auth service
