+++ api/tests/unit/domain/services/token-service_test.js
@@ -35,26 +35,6 @@ describe('Unit | Domain | Service | Token Service', function () {
     });
   });
 
-  describe('#createAccessTokenForCnav', function () {
-    it('should create access token with user id and source', function () {
-      // given
-      const userId = 123;
-      settings.authentication.secret = 'a secret';
-      settings.cnav.accessTokenLifespanMs = 1000;
-      const accessToken = 'valid access token';
-      const firstParameter = { user_id: userId, source: 'cnav' };
-      const secondParameter = 'a secret';
-      const thirdParameter = { expiresIn: 1 };
-      sinon.stub(jsonwebtoken, 'sign').withArgs(firstParameter, secondParameter, thirdParameter).returns(accessToken);
-
-      // when
-      const result = tokenService.createAccessTokenForCnav(userId);
-
-      // then
-      expect(result).to.be.deep.equal(accessToken);
-    });
-  });
-
   describe('#createIdTokenForUserReconciliation', function () {
     it('should return a valid idToken with firstName, lastName, samlId', function () {
       // given
